+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `mwc-ripple` now has CSS properties `--mdc-ripple-color`, `--mdc-ripple-fg-opacity`, and `--mdc-ripple-hover-opacity`
 - Added `RippleHandlers` to `mwc-ripple` to provide an easy integration point for calling ripple API.
 - Added `light` property to `mwc-ripple` to help style ripples on dark surfaces.
+- `mwc-select` can now select items by setting `mwc-select.value`.
 
 ### Changed